scoring: extract geometry grid builders into osh_scoring_geometry_runtime.c
mesh_geometry_to_grid / cyl_geometry_to_grid / zone_bin_index were defined in
osh_scoring_step.c though declared in osh_scoring_geometry_runtime_internal.h,
so a reader did not find them where the header implied. Move them (with their
shared axis_index helper) into a dedicated osh_scoring_geometry_runtime.c, and
osh_-prefix the three cross-TU symbols (osh_scoring_geometry_mesh_to_grid /
_cyl_to_grid / _zone_bin_index) to satisfy DEVELOPER.md §6.1. Call sites in
the step and point drivers are updated; no behaviour change.

Addresses the "geometry grid builders live in the wrong file" and the rest of
the §6.1 namespacing item in #251.
